Move load_ipython_extension into init, and defer loading magic, so that the magic_name could be modified before loading.
