+++ requirements.txt
@@ -1,5 +1,5 @@
 # development / testing only
-duckdb>=0.9.2
+duckdb>=1.0.0
 pandas>=2.1.2
 ipython>=8.0.0
 pyarrow>=12.0.0
